% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.mindist.block.R
\name{find.mindist.block}
\alias{find.mindist.block}
\title{Compute the minimum distance between data block.}
\usage{
find.mindist.block(dat, indx1, indx2, seg.length = 100, n.cores = 4)
}
\arguments{
\item{n.cores}{}
}
\value{
aa
}
\description{
Compute the minimum distance between data block.
}
