% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_k.R
\name{find_k}
\alias{find_k}
\title{Find \emph{k} function}
\usage{
find_k(dataset, train_size, iters, max_k)
}
\arguments{
\item{dataset}{Dataset from which to generate bootstrap.}

\item{iters}{Number of iterations/samples the bootstrap will take.}

\item{max_k}{Maximum \emph{k} value to analize.}
}
\value{
Returns a list containing:
\describe{
\item{\code{all_best_ks}}{A data.frame containing the best \emph{k} found for each clustering per sample.}
\item{\code{k_stats}}{A data.frame of the mean best \emph{k} and confidence intervals found for each clustering.}
\item{\code{best_k}}{A data.frame of the best \emph{k} found for each clustering.}
\item{\code{c_samples}}{A data.frame of the bootrapped data clustered with optimal \emph{k} s.}
}
}
\description{
A bootstrap method for calculating optimal \emph{k} s for each given clustering algorithm by:
\enumerate{
\item Sample of a given size from the dataset with replacement.
\item Calculate Dunn index for said sample per clustering algorithm for \emph{k} s between and 2 and a given value.
\item Find \emph{k} that maximizes the Dunn index per clustering algorithm.
\item Repeat steps 1 to 3 \emph{iters} times.
\item Get mean best \emph{k} per clustering algorithm.
}
}
\examples{
data   <- load("my_data.RData")
res    <- bootstrap(dataset = data, iters = 30, max_k = 5)
all_ks <- res$_all_best_ks
best_k <- res$best_k
k_st   <- res$k_stats
c_samp <- res$c_samples
}
