% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.r
\name{partition}
\alias{partition}
\title{Randomly partition a dataset}
\usage{
partition(dataset, train_size)
}
\arguments{
\item{dataset}{Dataset from which to generate partitions.}

\item{train_size}{Fraction of data to partition for training.}
}
\value{
Returns a list containing:
\describe{
\item{\code{train}}{The training data.frame partition.}
\item{\code{test}}{The testing data.frame partition.}
}
}
\description{
Randomly generate train and test partitions from a dataset.
}
\examples{
data  <- load("my_data.RData")
res   <- bootstrap(dataset = data, train_size = 0.1)
train <- res$train
test  <- res$test
}
