% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_clust_aov.R
\name{var_clust_aov}
\alias{var_clust_aov}
\title{Variable per clustering method ANOVA test}
\usage{
var_clust_aov(dataset, methods, dep_vars)
}
\arguments{
\item{dataset}{Dataset from which to perform test.}

\item{methods}{Character vector with the names of the columns corresponding to the clustering labels.}

\item{dep_vars}{Chracter vector with the names of the dependent variables to test.}
}
\value{
Returns a data.frame with the results of the ANOVA test.
}
\description{
Perform an ANOVA test to a chosen set of dependent variables with each of the given clusterings.
}
\examples{
a_o_v  <- var_clust_aov(dataset = data, methods = c("KMeans", "FCMeans", "HCWard", "HCSing", "HCComp"), 
dep_vars = c("WOMACPAIN", "BLKL", "TKREvent"))
}
