% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn.block.R
\name{dunn.block}
\alias{dunn.block}
\title{Compute Dunn index usign blockwise computation}
\usage{
dunn.block(
  clusters,
  Data = NULL,
  method = "euclidean",
  seg.length = min(100, nrow(Data)),
  n.cores = 4
)
}
\arguments{
\item{seg.length}{(=default=100)}
}
\value{
dunn
}
\description{
Compute Dunn index usign blockwise computation
}
\examples{
require(MASS)
n=1000
sigma=matrix(0.4, 2, 2);diag(sigma)=1
x = rbind(mvrnorm(n=n, rep(0,2),sigma),
mvrnorm(n=n, rep(5,2),sigma),
mvrnorm(n=n, rep(-5,2),sigma))
cfit = lapply(2:5,function(k)kmeans(x,k,nstart=10))
dunn.score<-unlist(mclapply(cfit, function(obj) dunn.block( obj$cluster,Data=x), mc.cores=4))
}
