% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.maxdist.block.R
\name{find.maxdist.block}
\alias{find.maxdist.block}
\title{Find the maximum distance between data block}
\usage{
find.maxdist.block(dat, indx, seg.length = 1000)
}
\arguments{
\item{seg.length}{}
}
\value{
aa
}
\description{
Find the maximum distance between data block
}
